ASoC: Intel: Fix to check CPA bit in DSP core power down
Skylake driver will set the SPA bit to 0 to turn of the
DSP core. Driver will poll the Current Power Active bit to
match the SPA bit value. when CPA bit matches the value of
SPA bit, the achieved power state has reached.

In case of DSP_power_down, register that was polled is SPA
instead of CPA. This patch corrects the register to be polled
in case of DSP power down.

Signed-off-by: Jeeja KP <jeeja.kp@intel.com>