ASoC: Intel: Fix not to handle interrupts if not for DSP
IRQ registered for Skylake driver is Shared. Interrupt
handlers needs to check the Interrupt status register
to identify if the interrupt has to be handled or not.

in case of DSP interrupts, the sst interrupt handler
will read the ADSPIS register and if the val is 1,
this indicates Audio DSP interrupt.

In case D3, DSP interrupt will be disabled and the ADSPIS
register value is set to 0xfffff. Added condition to check
the value of ADSPIS is 0xfffff, then return IRQ_NONE.

Signed-off-by: Jeeja KP <jeeja.kp@intel.com>