ASoC: topology: ABI - Add the types for HW configurations of BE/CC links
Define the types and ABI struct for a single BE or CC runtime supported
hardware config, e.g. audio hardware formats.

Topology provides this as a fallback if such HW settings are not available
in ACPI or device tree, to avoid hard code in drivers.

It's only for config items that can be programmed by SW or FW, not for
physical things like link connections or GPIO used for HP etc.

Signed-off-by: Mengdong Lin <mengdong.lin@linux.intel.com>