ASoC: topology: ABI - Add HW configurations to BE/CC links
Define the types and ABI struct for a single Backend or Codec<->Codec link
runtime supported hardware config, e.g. audio hardware formats.

The default HW config ID will help topology to find the DAI format to set
on init.

Topology provides this as a fallback if such HW settings are not available
in ACPI or device tree, to avoid hard code in drivers.

It's only for config items that can be programmed by SW or FW, not for
physical things like link connections or GPIO used for HP etc.