ASoC: SOF: Create HDMI DAI links for nocodec machine
If HDA support is enabled, this patch will create HDMI FE and BE
DAI links for the nocodec machine.

The 'cannot_ignore' flag of each HDMI FE link is set to 1, so they
won't be ignored during SoC card intialization. The DAI links coming
with platform topology file will coexist with them.

The number of HDMI links is hard coded to 3. Later, we can probe
the HDMI audio codec and check how many pins are enabled actually
and then decide the number of HDMI links.

Signed-off-by: Mengdong Lin <mengdong.lin@linux.intel.com>